git-svn-id: https://svn.st.cs.uni-sb.de/svn/st/projects/java-slicer/trunk@3702 ef2ef919-bb9b-4b18-8ff0-e5cc2bf85462

+++ java-slicer-tracer/src/main/java/de/unisb/cs/st/javaslicer/tracer/traceResult/TraceResult.java
@@ -119,7 +119,7 @@ public class TraceResult {
             if (threadToTrace == null) {
                 if (tracing == null)
                     tracing = t;
-            } else if (t.getThreadId() == threadToTrace)
+            } else if (t.getThreadId() == threadToTrace.longValue())
                 tracing = t;
             System.out.format("%15d: %s%n", t.getThreadId(), t.getThreadName());
         }
